--- src/main/java/org/spongepowered/common/mixin/core/item/inventory/MixinEmptyViewableCustomInventory.java
@@ -1,55 +1,0 @@
-/*
- * This file is part of Sponge, licensed under the MIT License (MIT).
- *
- * Copyright (c) SpongePowered <https://www.spongepowered.org>
- * Copyright (c) contributors
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package org.spongepowered.common.mixin.core.item.inventory;
-
-import org.spongepowered.api.item.inventory.Carrier;
-import org.spongepowered.api.item.inventory.custom.ContainerType;
-import org.spongepowered.api.item.inventory.type.CarriedInventory;
-import org.spongepowered.api.item.inventory.type.ViewableInventory;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.common.item.inventory.custom.EmptyViewableCustomInventory;
-import org.spongepowered.common.item.inventory.custom.ViewableCustomInventory;
-import org.spongepowered.common.mixin.core.inventory.MixinInventoryBasic;
-
-import java.util.Optional;
-
-@SuppressWarnings("rawtypes")
-@Mixin(EmptyViewableCustomInventory.class)
-public abstract class MixinEmptyViewableCustomInventory implements ViewableInventory, CarriedInventory<Carrier> {
-
-    @Shadow(remap = false) private Carrier carrier;
-    @Shadow(remap = false) private ContainerType type;
-
-    @Override
-    public ContainerType getType() {
-        return this.type;
-    }
-
-    @Override
-    public Optional<Carrier> getCarrier() {
-        return Optional.ofNullable(this.carrier);
-    }
-}
