+++ src/main/java/org/spongepowered/common/mixin/core/item/inventory/container/MixinCarrierContainerEnchantment.java
@@ -22,23 +22,34 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.common.mixin.core.item.inventory;
+package org.spongepowered.common.mixin.core.item.inventory.container;
 
-import org.spongepowered.api.item.inventory.custom.ContainerType;
-import org.spongepowered.api.item.inventory.type.ViewableInventory;
+import com.flowpowered.math.vector.Vector3d;
+import net.minecraft.inventory.ContainerEnchantment;
+import net.minecraft.util.math.BlockPos;
+import org.spongepowered.api.item.inventory.Carrier;
+import org.spongepowered.api.item.inventory.type.CarriedInventory;
+import org.spongepowered.api.world.Location;
+import org.spongepowered.api.world.World;
+import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.common.item.inventory.custom.ViewableCustomInventory;
+import org.spongepowered.common.interfaces.IMixinSingleBlockCarrier;
 
-@SuppressWarnings("rawtypes")
-@Mixin(ViewableCustomInventory.class)
-public abstract class MixinViewableCustomInventory extends MixinCustomInventory implements ViewableInventory {
+@Mixin(ContainerEnchantment.class)
+public class MixinCarrierContainerEnchantment implements IMixinSingleBlockCarrier {
 
-    @Shadow(remap = false) private ContainerType type;
+    @Shadow @Final public net.minecraft.world.World world;
+    @Shadow @Final public BlockPos position;
 
     @Override
-    public ContainerType getType() {
-        return this.type;
+    public Location getLocation() {
+        return new Location(((World) this.world), new Vector3d(this.position.getX(), this.position.getY(), this.position.getZ()));
     }
 
+    @SuppressWarnings({"rawtypes", "unchecked"})
+    @Override
+    public CarriedInventory<? extends Carrier> getInventory() {
+        return ((CarriedInventory) this);
+    }
 }
