+++ src/main/java/org/spongepowered/common/mixin/core/item/inventory/container/MixinSlot.java
@@ -22,7 +22,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.spongepowered.common.mixin.core.item.inventory;
+package org.spongepowered.common.mixin.core.item.inventory.container;
 
 import net.minecraft.inventory.IInventory;
 import net.minecraft.inventory.Slot;
