+++ src/main/resources/mixins.common.core.json
@@ -293,11 +293,11 @@
         "entity.projectile.MixinEntityWitherSkull",
         "event.cause.entity.damage.MixinMinecraftBlockDamageSource",
         "event.cause.entity.damage.MixinMinecraftFallingBlockDamageSource",
-        "inventory.MixinInventoryBasic",
-        "inventory.MixinInventoryCrafting",
-        "inventory.MixinInventoryEnderChest",
-        "inventory.MixinInventoryHelper",
-        "inventory.MixinInventoryLargeChest",
+        "item.inventory.vanilla.MixinInventoryBasic",
+        "item.inventory.vanilla.MixinInventoryCrafting",
+        "item.inventory.vanilla.MixinInventoryEnderChest",
+        "item.inventory.MixinInventoryHelper",
+        "item.inventory.vanilla.MixinInventoryLargeChest",
         "item.MixinEnchantment",
         "item.MixinItem",
         "item.MixinItemArmorMaterial",
@@ -311,26 +311,26 @@
         "item.data.MixinItemPotion",
         "item.data.MixinItemWritableBook",
         "item.data.MixinItemWrittenBook",
-        "item.inventory.MixinCarrierContainerEnchantment",
-        "item.inventory.MixinCarrierContainerRepair",
-        "item.inventory.MixinCarrierContainerWorkbench",
-        "item.inventory.MixinContainer",
-        "item.inventory.MixinContainerCanInteract",
-        "item.inventory.MixinContainerCrafting",
-        "item.inventory.MixinContainerHorseChest",
-        "item.inventory.MixinContainerPlayer",
-        "item.inventory.MixinContainerRepair",
-        "item.inventory.MixinContainerWorkbench",
-        "item.inventory.MixinCustomInventory",
-        "item.inventory.MixinCustomContainer",
-        "item.inventory.MixinEmptyViewableCustomInventory",
-        "item.inventory.MixinEntityCarriedInventory",
+        "item.inventory.container.MixinCarrierContainerEnchantment",
+        "item.inventory.container.MixinCarrierContainerRepair",
+        "item.inventory.container.MixinCarrierContainerWorkbench",
+        "item.inventory.container.MixinContainer",
+        "item.inventory.container.MixinContainerCanInteract",
+        "item.inventory.container.MixinContainerCrafting",
+        "item.inventory.container.MixinContainerHorseChest",
+        "item.inventory.container.MixinContainerPlayer",
+        "item.inventory.container.MixinContainerRepair",
+        "item.inventory.container.MixinContainerWorkbench",
+        "item.inventory.custom.MixinCustomInventory",
+        "item.inventory.custom.MixinCustomContainer",
+        "item.inventory.custom.MixinEmptyViewableCustomInventory",
+        "item.inventory.vanilla.MixinEntityCarriedInventory",
         "item.inventory.MixinTraitViewable",
         "item.inventory.MixinTraitInventory",
-        "item.inventory.MixinTraitAdapter",
-        "item.inventory.MixinSlot",
-        "item.inventory.MixinSlotCrafting",
-        "item.inventory.MixinViewableCustomInventory",
+        "item.inventory.vanilla.MixinTraitAdapter",
+        "item.inventory.container.MixinSlot",
+        "item.inventory.container.MixinSlotCrafting",
+        "item.inventory.custom.MixinViewableCustomInventory",
         "item.recipe.crafting.MixinIngredient",
         "item.recipe.crafting.MixinIRecipe",
         "item.recipe.crafting.MixinShapedRecipes",
